# [core] util::tileCover optimization: three scans and no duplicates handling.