@astojilj 2019-12-03
Closing in favor of https://github.com/mapbox/mapbox-gl-js/pull/8975. @asheemmamoowala, thanks for review.